More fiddling with the test interface
I have an idea in mind. What I'd like to accomplish is to have rqc run
each test target, forking in the manner of AFL, and communicating
facts to the runner from the target via an mmap. More ambitious,
expose an interface to the test writer allowing them to introduce
limited telemetry about the run. Ambition is to allow the test runner
to manipulate its own byte buffer according to some score, also in the
manner of a fuzzer.

Might dump the stdin interface, opting entirely for mmap. Unsure.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>