Able to run tests, kind of
In this commit RQC is now able to run tests, sort of. The 'test' is
just a random byte blob that the client interprets, so far so
good. But, the server is blind as to the result and has no shrinking
capability. The protocol for notification is also wrong.

Buuut, it's a start.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>