Now able to run tests, kind of
With this commit we've got the basic protocol for the client and
server down. We're now able to run test targets, restarting the target
if it cleanly exits and are recording the test results from the
runs. I noticed that I'm creating zombies, which is a no-no, so that's
something that needs to be cleaned up. Also, still no shrinking yet.